backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1

pydantic==2.9.2
python-dotenv==1.0.1

openai==1.51.0
requests==2.32.3

# Optionnel : si tu utilises OR-Tools pour un CSP avancé
ortools==9.14.6206

# Tests unitaires
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.28.1
